+++ lib/devise/jwt/cookie2.rb
